Precalculate Script.sha when creating scripts (#51)
This is a workaround for a problem in redis-py.

Script objects are initialized with sha set to the empty string, which means that the first time the Script is executed, it makes an EVALSHA call that is guaranteed to fail. This produces extra traffic for no apparent benefit. A fix has been submitted upstream, but this works around it in the meantime.